requirements.txt
absl-py==1.4.0
anyio==3.6.2
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
exceptiongroup==1.1.1
fastapi==0.95.1
Flask==2.2.3
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
idna==3.4
image==1.5.33
iniconfig==2.0.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
MyApplication==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pytest==7.3.1
python-multipart==0.0.6
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.26.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
wrapt==1.14.1
